pyproject.toml
[tool.poetry]
name = "qaas"
version = "0.1.0"
description = "Create and participate in quizzes"
authors = ["Igor Klepuszewski <igor.klepuszewski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.1.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
django-invitations = "^2.0.0"
django-filter = "^23.2"
pytest = "^7.4.0"
model-bakery = "^1.12.0"
freezegun = "^1.2.2"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
